+++ packages/client/CHANGELOG.md
@@ -21,6 +21,7 @@
 - BREAKING: `getBody` and `getHeader` only accept block-hashes.
 - BREAKING: renamed `watchBlockBody` to `getBlockBody$`
 - BREAKING: `watchValue` options are now consistent with `watchEntries` options, and also with the rest of query options.
+- BREAKING: rename chainspecs `ksmcc3` => `kusama`, `westend2` => `westend`
 
 - **WS-Provider:**
   - BREAKING: This export has been moved in favor of `polkadot-api/ws`